feat: adjust library designs
