+++ src/index.ts
@@ -1,5 +1,3 @@
-export { SoftDeletable } from "./soft-deletable.decorator";
-export { SOFT_DELETABLE } from "./soft-deletable.symbol";
-export { SOFT_DELETABLE_FILTER } from "./soft-deletable-filter.constant";
-export { SoftDeletableHandlerSubscriber } from "./soft-deletable-handler.subscriber";
-export { SoftDeletableMetadata } from "./soft-deletable-metadata.interface";
+export * from "./common";
+export * from "./decorator";
+export * from "./handler";
