--- src/soft-deletable-filter.constant.ts
@@ -1,1 +1,0 @@
-export const SOFT_DELETABLE_FILTER = "soft-deletable-filter";
