--- src/soft-deletable-metadata.interface.ts
@@ -1,4 +1,0 @@
-export interface SoftDeletableMetadata<Entity, Field extends keyof Entity> {
-  field: Field;
-  value: () => Entity[Field];
-}
