--- src/soft-deletable.symbol.ts
@@ -1,1 +1,0 @@
-export const SOFT_DELETABLE = Symbol("soft-deletable");
